+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Bug fix: The one instance functionality was executed after the update check
 - Bug fix: Problems in the installation Tab when uninstall->reinstall Games
 - Bug fix: New users do not show on the community page without client restart
+- Bug fix: Added a few more safeties in the offline cache
 
 ## 1.6.2
 ### Changes